refactor: move all profiler metrics into context
